+++ tests/_data/StubProject/src/Controller/Index.php
@@ -38,4 +38,9 @@ class Index extends Controller
             )
         );
     }
+
+    public function exceptionAction()
+    {
+        throw new \Exception('user error');
+    }
 }