+++ tests/_data/StubProject/src/Task/Custom.php
@@ -4,24 +4,36 @@ namespace StubProject\Task;
 
 use Phalcon\Cli\Task;
 
+/**
+ * @property \Phalcon\Cli\Dispatcher $dispatcher
+ */
 class Custom extends Task
 {
-    function mainAction()
+    public function mainAction()
     {
         echo __CLASS__ . '::' . __FUNCTION__ . '()';
     }
 
-    function customAction()
+    public function customAction()
     {
         echo __CLASS__ . '::' . __FUNCTION__ . '()';
     }
-    
-    function argumentAction(array $arguments)
+
+    public function argumentAction(array $arguments)
     {
-        echo __CLASS__ . '::' . __FUNCTION__ . '(' . implode(', ', $arguments). ')';
+        echo __CLASS__ . '::' . __FUNCTION__ . '(' . implode(', ', $arguments) . ')';
     }
 
-    function serviceAction()
+    public function optionAction(array $arguments)
+    {
+        $opts = [];
+        foreach ($this->dispatcher->getOptions() as $key => $value) {
+            $opts[] = '--' . $key . '=' . $value;
+        }
+        echo __CLASS__ . '::' . __FUNCTION__ . '(' . implode(', ', $arguments) . ', '. implode(', ', $opts). ')';
+    }
+
+    public function serviceAction()
     {
         if (false === $this->getDI()->get('config')['customKey']) {
             throw new \Exception;