+++ tests/functional/CommandLineCest.php
@@ -8,14 +8,18 @@ class CommandLineCest
             'taskDefaultNamespace' => 'StubProject\Task',
         ],
         'services'   => [
-            'stdClass'  => 'StubProject\Service\StdClass',
-            'stdClass2' => 'StubProject\Service\StdClass2',
+            'stdClass'     => 'StubProject\Service\StdClass',
+            'stdClass2'    => 'StubProject\Service\StdClass2',
+            'errorHandler' => 'StubProject\Service\ErrorHandler',
         ],
         'customKey'  => true,
     ];
 
     /** @var array */
     private $arguments;
+
+    /** @var string */
+    private $forcedExpection;
 
     /**
      * @param FunctionalTester $tester
@@ -37,20 +41,29 @@ class CommandLineCest
 
         $tester->comment('`> php index.php ' . implode(' ', $helpDoc) . '`');
 
-        $expected = $this->config['dispatcher']['taskDefaultNamespace'] . '\\'
-            . ucfirst($task) . '::' . $action . 'Action('
-            . implode(', ', $arguments)
-            . ')';
+        if ($this->forcedExpection) {
+            $expected = $this->forcedExpection;
+        } else {
+            $expected = $this->config['dispatcher']['taskDefaultNamespace'] . '\\'
+                . ucfirst($task) . '::' . $action . 'Action('
+                . implode(', ', $arguments)
+                . ')';
+        }
         
         ob_start();
         $tester->execute(function () {
-            \Phapp\Application\Bootstrap::init($this->config)->runApplicationOn(
-                ['argv' => array_merge([__FILE__], $this->arguments)]
-            );
+            try {
+                \Phapp\Application\Bootstrap::init($this->config)->runApplicationOn(
+                    ['argv' => array_merge([__FILE__], $this->arguments)]
+                );
+            } catch (\Throwable $e) {
+                echo $e->getMessage();
+            }
         });
         $tester->assertSame($expected, ob_get_clean());
 
         $this->arguments = [];
+        $this->forcedExpection = null;
     }
 
     /**
@@ -115,5 +128,23 @@ class CommandLineCest
     public function callCustomActionInCustomTaskWithServiceAccess(FunctionalTester $tester)
     {
         $this->arguments = ['custom', 'service'];
+    }
+
+    /**
+     * @param FunctionalTester $tester
+     */
+    public function callCustomActionWichThrowsAnException(FunctionalTester $tester)
+    {
+        $this->forcedExpection = 'error';
+        $this->arguments = ['main', 'error'];
+    }
+
+
+    /**
+     * @param FunctionalTester $tester
+     */
+    public function callCustomActionInCustomTaskWithNormalizedArguments(FunctionalTester $tester)
+    {
+        $this->arguments = ['custom', 'option', 'arg', '--first=12', '--second=foo'];
     }
 }